Fixing a subtle bug in the lookup of already-used-charsets for the translation from UTF-8 to MARC8

+++ src/org/marc4j/converter/impl/ReverseCodeTable.java
@@ -244,7 +244,7 @@ public abstract class ReverseCodeTable
         for (int i = 0; i < charsetsUsed.length(); i++)
         {
             char toUse = charsetsUsed.charAt(i);
-            if (chars.containsKey(toUse))
+            if (chars.containsKey((int)toUse))
             {
                 return(toUse);
             }
